tools/x-bot/requirements.txt
# X/Twitter Bot Dependencies
tweepy>=4.14.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
schedule>=1.2.0
openai>=1.0.0
rich>=13.0.0
httpx>=0.25.0
